Incremental Refactor Cleanup to make Biometricks.kt smaller.
