+++ biometricks/src/main/java/com/handstandsam/biometricks/Biometricks.kt
@@ -1,20 +1,14 @@
 package com.handstandsam.biometricks
 
-import android.app.Activity
 import android.content.Context
-import android.os.Build
-import android.os.Handler
-import android.os.Looper
-import android.view.ViewTreeObserver
 import androidx.annotation.MainThread
 import androidx.biometric.BiometricPrompt
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
+import com.handstandsam.biometricks.internal.AuthenticationCallbackWrapper
 import com.handstandsam.biometricks.internal.BiometricksHelper
-import kotlin.coroutines.Continuation
-import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
+import com.handstandsam.biometricks.internal.UiHelpers
 import kotlin.coroutines.suspendCoroutine
 
 /**
@@ -111,14 +105,14 @@ sealed class Biometricks {
             promptInfo: BiometricPrompt.PromptInfo,
             showLoading: (Boolean) -> Unit
         ): BiometricPrompt.CryptoObject {
-            ensureFocus(activity)
+            UiHelpers.ensureFocus(activity)
 
-            return handleApi28Loading(activity, showLoading) {
+            return UiHelpers.handleApi28Loading(activity, showLoading) {
                 suspendCoroutine<BiometricPrompt.CryptoObject> { continuation ->
                     BiometricPrompt(
                         activity,
                         ContextCompat.getMainExecutor(activity),
-                        Callback(continuation)
+                        AuthenticationCallbackWrapper(continuation)
                     ).authenticate(promptInfo, cryptoObject)
                 }
             }
@@ -168,129 +162,9 @@ sealed class Biometricks {
             showLoading: (Boolean) -> Unit
         ): BiometricPrompt.CryptoObject {
             val activity = fragment.requireActivity()
-
-            ensureFocus(activity)
-
-            return handleApi28Loading(activity, showLoading) {
-                suspendCoroutine<BiometricPrompt.CryptoObject> { continuation ->
-                    BiometricPrompt(
-                        fragment,
-                        ContextCompat.getMainExecutor(fragment.requireContext()),
-                        Callback(continuation)
-                    ).authenticate(promptInfo, cryptoObject)
-                }
-            }
+            return showPrompt(activity, cryptoObject, promptInfo, showLoading)
         }
 
-        private suspend fun <T> handleApi28Loading(
-            activity: Activity,
-            showLoading: (Boolean) -> Unit,
-            showPrompt: suspend () -> T
-        ): T {
-            // On api 28 if the user is locked out of biometrics from too many failed
-            // attempts there will be a long delay before getting the error back. So the
-            // user isn't confused as to what is going on, show a loading indicator.
-            if (Build.VERSION.SDK_INT == 28) {
-                val window = activity.window
-                val handler = Handler(Looper.getMainLooper())
 
-                var showingLoading = false
-
-                // Don't show if the prompt shows quickly
-                val showLoadingRunnable = Runnable {
-                    showingLoading = true
-                    showLoading(true)
-                }
-                handler.postDelayed(showLoadingRunnable, 150)
-
-                // The only way to tell that the prompt is shown is to listen to the window
-                // losing focus events.
-                window.decorView.viewTreeObserver.addOnWindowFocusChangeListener(object :
-                    ViewTreeObserver.OnWindowFocusChangeListener {
-                    override fun onWindowFocusChanged(hasFocus: Boolean) {
-                        if (!hasFocus) {
-                            window.decorView.viewTreeObserver.removeOnWindowFocusChangeListener(this)
-                            handler.removeCallbacks(showLoadingRunnable)
-                            showingLoading = false
-                            showLoading(false)
-                        }
-                    }
-                })
-
-                try {
-                    return showPrompt()
-                } finally {
-                    handler.removeCallbacks(showLoadingRunnable)
-                    // Hide loading if the prompt fails to show
-                    if (showingLoading) {
-                        showLoading(false)
-                    }
-                }
-            } else {
-                // The prompt shows immediately on api 29+ so don't worry about any of this.
-                return showPrompt()
-            }
-        }
-
-        private suspend fun ensureFocus(activity: Activity) {
-            suspendCoroutine<Unit> { continuation ->
-                // Showing the biometrics prompt will be ignored if the app does not have focus. You may
-                // think that this will always be the case if you are resumed, it is not.
-                if (activity.hasWindowFocus()) {
-                    continuation.resume(Unit)
-                } else {
-                    val window = activity.window
-                    window.decorView.viewTreeObserver.addOnWindowFocusChangeListener(object :
-                        ViewTreeObserver.OnWindowFocusChangeListener {
-                        override fun onWindowFocusChanged(hasFocus: Boolean) {
-                            if (hasFocus) {
-                                window.decorView.viewTreeObserver.removeOnWindowFocusChangeListener(
-                                    this
-                                )
-                                continuation.resume(Unit)
-                            }
-                        }
-                    })
-                }
-            }
-        }
-
-        private class Callback(private val continuation: Continuation<BiometricPrompt.CryptoObject>) :
-            BiometricPrompt.AuthenticationCallback() {
-            private var authenticationFailed = false
-
-            override fun onAuthenticationError(
-                errorCode: Int,
-                errString: CharSequence
-            ) {
-                val shouldShow =
-                    !isCancel(errorCode) && !authenticationFailed
-                continuation.resumeWithException(
-                    BiometricException(
-                        errorCode,
-                        errString,
-                        shouldShow
-                    )
-                )
-            }
-
-            override fun onAuthenticationSucceeded(result: BiometricPrompt.AuthenticationResult) {
-                continuation.resume(result.cryptoObject!!)
-            }
-
-            override fun onAuthenticationFailed() {
-                // This means the dialog was shown, so we don't want to show the
-                // error again ourselves
-                authenticationFailed = true
-            }
-
-            /**
-             * If the prompt was canceled we don't want to show an error ourselves
-             */
-            fun isCancel(errorCode: Int) =
-                errorCode == BiometricPrompt.ERROR_CANCELED
-                        || errorCode == BiometricPrompt.ERROR_USER_CANCELED
-                        || errorCode == BiometricPrompt.ERROR_NEGATIVE_BUTTON
-        }
     }
 }
